local-open-interpreter/pyproject.toml
[tool.poetry]
name = "local-open-interpreter"
version = "0.1.0"
description = "Run a local Code Interpreter"
authors = ["Daniel Lauzon <daniel.lauzon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
open-interpreter = "^0.1.15"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
